moves the command
